modelserver/requirements.txt
redis
scikit-image
pillow
numpy
ipdb
jupyter
ipython
opencv-python
pandas
python-multipart